feat: Light cleanup.

+++ src/constants/src/title_protocol.rs
@@ -1,8 +1,8 @@
 pub mod title_protocol {
-    pub const INIT_GAME_CONNECTION: i32 = 14;
+    pub const INIT_GAME_CONNECTION: u8 = 14;
     pub const INIT_JS5REMOTE_CONNECTION: u8 = 15;
-    pub const LOGIN: i32 = 16;
-    pub const RECONNECT: i32 = 18;
+    pub const LOGIN: u8 = 16;
+    pub const RECONNECT: u8 = 18;
     
     pub const CREATE_CHECK_DATE_COUNTRY: i32 = 147;
     pub const CREATE_CHECK_NAME: i32 = 186;
